refactor: relocate react examples
