+++ articles/components/accordion/index.asciidoc
@@ -30,7 +30,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/accordion/AccordionBasic
 
 [source,tsx]
 ----
-include::{root}/frontend/demo/react/accordion/accordion-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/accordion/react/accordion-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 --
