+++ articles/components/avatar/index.asciidoc
@@ -28,7 +28,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/avatar/AvatarBasic.java[
 
 [source,tsx]
 ----
-include::{root}/frontend/demo/react/avatar/avatar-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/avatar/react/avatar-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 
 --
