+++ articles/components/badge/index.asciidoc
@@ -26,7 +26,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/badge/BadgeBasic.java[re
 
 [source,tsx]
 ----
-include::{root}/frontend/demo/react/badge/badge-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/badge/react/badge-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 
 --
