+++ articles/components/button/index.asciidoc
@@ -31,7 +31,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/button/ButtonBasic.java[
 
 [source,tsx]
 ----
-include::{root}/frontend/demo/react/button/button-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/button/react/button-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 
 --
