+++ articles/components/combo-box/index.asciidoc
@@ -40,7 +40,7 @@ include::{root}/src/main/java/com/vaadin/demo/domain/Country.java[group=Java,tag
 
 [source,tsx]
 ----
-include::{root}/frontend/demo/react/combobox/combo-box-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/combobox/react/combo-box-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 --
