+++ articles/components/grid/index.asciidoc
@@ -42,7 +42,7 @@ include::{root}/src/main/java/com/vaadin/demo/domain/Person.java[group=Java,tags
 
 [source,typescript]
 ----
-include::{root}/frontend/demo/react/grid/grid-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/grid/react/grid-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 
 --
@@ -808,7 +808,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/grid/GridStyling.java[re
 
 [source,typescript]
 ----
-include::{root}/frontend/demo/react/grid/grid-styling.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/grid/react/grid-styling.tsx[render,tags=snippet,indent=0,group=React]
 ----
 
 [source,css]
