+++ articles/components/login/index.asciidoc
@@ -39,7 +39,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/login/LoginBasic.java[re
 
 [source,tsx]
 ----
-include::{root}/frontend/demo/react/login/login-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/login/react/login-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 --
