+++ articles/components/notification/index.asciidoc
@@ -61,7 +61,7 @@ include::{root}/src/main/java/com/vaadin/demo/component/notification/Notificatio
 
 [source,typescript]
 ----
-include::{root}/frontend/demo/react/notification/notification-basic.tsx[render,tags=snippet,indent=0,group=React]
+include::{root}/frontend/demo/component/notification/react/notification-basic.tsx[render,tags=snippet,indent=0,group=React]
 ----
 --
