+++ frontend/demo/component/app-layout/react/app-layout-basic.tsx
@@ -1,4 +1,4 @@
-import { reactExample } from 'Frontend/demo/react/react-example'; // hidden-source-line
+import { reactExample } from 'Frontend/demo/react-example'; // hidden-source-line
 export default reactExample(Example); // hidden-source-line
 import React from 'react';
 import { AppLayout } from '@hilla/react-components/AppLayout.js';
