+++ frontend/demo/component/button/react/button-basic.tsx
@@ -1,4 +1,4 @@
-import { reactExample } from 'Frontend/demo/react/react-example'; // hidden-source-line
+import { reactExample } from 'Frontend/demo/react-example'; // hidden-source-line
 export default reactExample(Example); // hidden-source-line
 import React, { useState } from 'react';
 import { Button } from '@hilla/react-components/Button.js';
